fix exception source by removing playsound for now

+++ funscript_editor/ui/funscript_generator_window.py
@@ -106,11 +106,13 @@ class FunscriptGeneratorWindow(QtWidgets.QMainWindow):
 
             self.__logger.info("Save result to %s", self.output_file)
             if not success: self.__show_message(msg, error=True)
-            if platform.system() == "Windows":
-                # On windows i have problems to close the process
-                # Workaround we kill with taskkill
-                os.system('taskkill /f /im funscript-editor.exe')
-            os._exit(os.EX_OK)
+            if False:
+                if platform.system() == "Windows":
+                    # On windows i have problems to close the process
+                    # Workaround we kill with taskkill
+                    os.system('taskkill /f /im funscript-editor.exe')
+                os._exit(os.EX_OK)
+            sys.exit()
 
 
     def run(self) -> None:
